Rename LEGACY_PurchasePage/AddForm
