+++ app/components/views/TicketsPage/PurchaseTab/LEGACY_PurchasePage/LEGACY_StakePools/StakePoolsList.jsx
@@ -126,10 +126,4 @@ const StakePoolsList = ({
   </>
 );
 
-StakePoolsList.propTypes = {
-  configuredStakePools: PropTypes.array.isRequired,
-  unconfiguredStakePools: PropTypes.array.isRequired,
-  onShowAddStakePool: PropTypes.func.isRequired
-};
-
 export default StakePoolsList;
